refactor to put deadletter message definitions in the types package
